requirements.txt
fastapi
uvicorn
pyodbc
pandas
python-dotenv